torchic/pyproject.toml
[project]
name = "torchic"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Hari Krishnan", email = "codeharik@gmail.com" }
]
dependencies = [
    "uvicorn>=0.30.6",
    "fastapi>=0.114.2",
    "torch>=2.4.1",
    "transformers>=4.44.2",
    "pillow>=10.4.0",
    "python-multipart>=0.0.9",
    "torchvision>=0.19.1",
    "requests>=2.32.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/torchic"]
